fix: move e2e tests to a single test script.
