+++ tests/e2e/000-npm-link.ts
@@ -1,0 +1,8 @@
+import {join} from 'path';
+import {npm} from './utils';
+
+export default function () {
+  // Setup to use the local angular-cli copy.
+  process.chdir(join(__dirname, '..'));
+  return npm('link');
+}
