+++ tests/e2e_workflow.ts
@@ -16,53 +16,19 @@ const http = require('http');
 const request = require('request');
 
 
-function isMobileTest() {
-  return !!process.env['MOBILE_TEST'];
-}
+/** Load all the files from the e2e, filter and sort them and build a promise of their default export. */
+fs.readdirSync(path.join(__dirname, 'e2e'))
+    .filter(name => name.match(/^\d\d\d/))
+    .sort()
+    .reduce((previous: Promise<void>, fileName: string) => {
+        return previous.then(() => {
+          console.log(chalk.green(`Running "${fileName}"...`));
+          return require(`./e2e/${fileName.replace(/\.ts$/)}`);
+        });
+      },
+      Promise.resolve());
 
-
-function execOrFail(cmd: string, ...args: string[]): Promise<string> {
-  let stdout = '';
-  const cwd = process.cwd();
-  console.log(chalk.yellow(stripIndents`
-    Running "${cmd} ${args.join(' ')}"...
-    CWD: ${cwd}
-  `));
-
-  const npmProcess = spawn(cmd, args, { cwd, detached: true });
-  npmProcess.stdout.on('data', (data: Buffer) => {
-    const str = data.toString('utf-8');
-    stdout += str;
-    process.stdout.write(chalk.reset('   ' + str.split(/\n/g).join('\n   ').replace(/ +$/, '')));
-  });
-  // npmProcess.stderr.on('data', (data: Buffer) => {
-    // process.stderr.write(chalk.red('   ' + data.toString().split(/\n/g).join('\n   ')));
-  // });
-
-  return new Promise((resolve, reject) => {
-    npmProcess.on('close', (code: number) => {
-      if (code == 0) {
-        resolve(stdout);
-      } else {
-        reject(new Error(oneLine`
-          Running "${cmd} ${args.join(' ')}" returned error code ${code}.
-        `));
-      }
-    });
-  });
-}
-
-
-function ng(...args: string[]) {
-  return execOrFail('ng', ...args)
-}
-
-
-function npm(...args: string[]) {
-  return execOrFail('npm', ...args);
-}
-
-
+/*
 Promise.resolve()
   .then(() => {
     // Setup to use the local angular-cli copy.
